Connect Router to Responders
I've configured the router so that it acts just like the event
sender-reciever we built (command pattern).  The router takes
a list of responders in its constructor, when get responses is called it
calls get response on each responder.

Next I plan to introduce a few example types of requests and responders
and then develop the logic for routing each request to the appropriate
responders.