Merge Router into RequestHandler
I've removed the concrete Router class from RequestRouter and replaced
it with an abstract class called RequestHandler.

The idea is that the RequestHandler base class is implemented for each
bespoke request type.  It is responsable for deciding which responders
are sent the request.  It is also responsable for converting the request
into the "Standard" request format and converting the "Standard"
responses it gets back into the format the requester expects.